add view & temlplates for expend, delete expend, refactoring,
