+++ moneta/etc/moneta_mysql.cnf
@@ -1,5 +1,5 @@
 [client]
 database = db_moneta
 user = moneta_user
-password = db_password
+password = DeathRow$1991
 default-character-set = utf8
