+++ moneta/www/settings/urls.py
@@ -1,22 +1,10 @@
-"""settings URL Configuration
-
-The `urlpatterns` list routes URLs to views. For more information please see:
-    https://docs.djangoproject.com/en/2.1/topics/http/urls/
-Examples:
-Function views
-    1. Add an import:  from my_app import views
-    2. Add a URL to urlpatterns:  path('', views.home, name='home')
-Class-based views
-    1. Add an import:  from other_app.views import Home
-    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
-Including another URLconf
-    1. Import the include() function: from django.urls import include, path
-    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
-"""
 from django.urls import path
 
+from ..views.expend import show_form, success_edit
+
+
 urlpatterns = [
-    path('expend/edit/', views.show_form),
-    path('expend/edit/success/', views.success_edit),
+    path('expend/edit/', show_form),
+    path('expend/edit/success/', success_edit),
 
 ]
