+++ moneta/www/templates/base.html
@@ -1,0 +1,45 @@
+{% load static %}
+
+<html lang="en">
+
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0/css/bootstrap.min.css"
+          integrity="sha384-Gn5384xqQ1aoWXA+058RXPxPg6fy4IWvTNh0E263XmFcJlSAwiGgFAW/dAiS6JXm" crossorigin="anonymous">
+    <link rel="stylesheet" href="{% static 'css/main.css' %}" type="text/css">
+</head>
+
+
+<body>
+<header class="header">
+    <div>
+        <a href="/">Moneta</a>
+    </div>
+
+    <div class="container-header">
+        {% if user.is_authenticated %}
+        <div class="header-label">
+            <a>{{ user.username }}</a>
+        </div>
+        <div class="header-label">
+            <a href="{% url 'logout' %}">Logout</a>
+        </div>
+        {% endif %}
+    </div>
+</header>
+
+<main role="main" class="container">
+    <div class>
+        <div>
+            {% block content %}
+            {% endblock %}
+        </div>
+    </div>
+</main>
+
+<script src="https://code.jquery.com/jquery-3.2.1.slim.min.js"
+        integrity="sha384-KJ3o2DKtIkvYIK3UENzmM7KCkRr/rE9/Qpg6aAZGJwFDMVNA/GpGFF93hXpG5KkN"
+        crossorigin="anonymous"></script>
+</body>
+</html>
