+++ moneta/www/templates/expend/expend_detailed.html
@@ -1,0 +1,14 @@
+{% extends "base.html" %}
+{% load static %}
+
+{% block title %}Expends{% endblock %}
+
+{% block content %}
+<h2>Details for {{expend_detailed.name}}</h2>
+
+<p> Cerrency: {{expend_detailed.currency}}</p>
+<p> Planned cost: {{expend_detailed.planned_cost}}</p>
+
+<a href="edit/">Edit</a>
+
+{% endblock content %}
