+++ moneta/www/templates/expend/success.html
@@ -1,0 +1,4 @@
+{% extends "base.html" %}
+{% block content %}
+    <h2>Your expand was successfully updated</h2>
+{% endblock content %}
