+++ moneta/www/templates/login_app/login.html
@@ -1,0 +1,18 @@
+{% extends "base.html" %}
+{% load crispy_forms_tags %}
+{% block content %}
+<div class="container">
+    <form method="POST">
+        {% csrf_token %}
+        <fieldset class="form-group">
+            <input type="hidden" id="csrf_token" value="{{ csrf_token }}">
+            {{ form|crispy }}
+        </fieldset>
+        <div>
+            <input type="submit" value="Log in"/>
+            <a href="#">Lost your password?</a>
+            <a href="#">Register</a>
+        </div>
+    </form>
+</div>
+{% endblock content %}
